I20_1-171 : TurboXAS : Adjust calculated step size to match resolution of encoder
Encoder resolution of the 'turbo slit x' motor is limited to 1e-4mm, and
sometimes a greater precision is needed to achieve requested number of points
for a scan. This can lead to fewer encoder pulses being produced per spectrum
than expected, and spectra appearing to shift in energy.
To help avoid this problem, the step size calculation now rounds the step
size down to the neared 1e-4 mm to match the encoder resolution.

Change-Id: I82620c5316e490fd7a9666f366a25f79d376dbf1