I20_1-172 : TurboXAS : Append points to trajectory scan profile when creating large scans
To send large trajectory scan profiles to Epics, it is necessary to send
the data in chunks by first 'building' a profile of at least 1000 points
and then 'appending' the rest of the points as necessary.
GDA now sends the trajectory scan point data by building and appending, so
that scans with large numbers of points can be done. Successfully created
and started scans with around 800,000 points. 

Unit tests have been updated to test trajectory build and append operations.

Change-Id: I7259840dbf0455dcdc1bd6d4e07d9dc77099447e