DAQ-1123 Remove i15-1 related errors from other beamline server
This problem was occurring because XpdfTaskConsumer and XpdfTaskRunner
were OSGi components which used the IXpdfDatabaseService. When they were
started they required the IXpdfDatabaseService which tried to open a
database connection to ISPyB. If the correct properties were not set
(which was the case for all beamlines except i15-1) this would fail
causing the exception.

The solution here is to create the XpdfTaskConsumer and XpdfTaskRunner
in Spring only on i15-1 so the IXpdfDatabaseService will not be required
on other beamlines and the error won't occur.

As Spring can't inject OSGi services the solution used is to obtain the
required services from the bundle activator. However the Jython server
can now be injected by Spring.

Also:
- Improve startup logging
- Update JavaDocs
- Remove unneeded dependencies from i15-1 server bundle
- Disable DS annotation builder as there are no longer any annotated
  OSGi components

Change-Id: I3c8e1ca52008e0d46a30f6d21f41ea9cf145a3a7