[DAQ-1130] Config changes to pass file prefix separately to IcatXMLCreator
Previously, the "directory" parameter of IcatXMLCreator also included
the prefix of the files to be written. This has now been changed,
so that the file prefix is passed separately and the output path
built up in a more robust way.

Also, remove directory from DummyIcatXMLCreator, as it is not used.

Change-Id: I314432850e213e1b01efb19f8fe068f381be512a