Add new plugin requirement org.dawnsci.processing.python to b16 client
Commit d98bc97 in scisoft-spectroscopy.git added org.dawnsci.processing.python
as an additional dependency of uk.ac.diamond.scisoft.spectroscopy. Therefore, all
beamlines that include u.a.d.s.spectroscopy also need to include o.d.p.python.
Without that, the client export fails.

Note that this is probably not the correct way to solve this. A better solution
would be to have a spectroscopy feature that can be added to beamline products
when required; however that is a significant change that can be considered later.