I20_1-193 - Linear/Cyclic experiment view : remove graphical summary of experiment
Hide the time bar section of the EDE 'time resolved experiment' gui. This is
done by setting the weight for that section to zero. It would be preferable
to not create that part of the composite at all but it seems that the timing
group section needs the JFace binding to it for the gui to work.

Change-Id: Idd8c9cc8bf1e1c30c913f82cc28da8c38a1cb1ec