DAQ-1037 I07 Fix plugin_customization
Delete the previously used file in i07-config/client

Put the contents of this in the plugin_customization.ini in the client
uk.ac.gda.beamline.i07 plugin. This is the only file that will now be
used.

If it is later desired to use an ini in the config project (as before)
then the values should be copied to the new location and the
relevant -D argument added to both the .product files.