I20_1-194 : Create Ascii output file for TurboXas scan.
A new AsciiWriter class has been added which can create an Ascii file from
the Nexus file produced by TurboXas scans.
- Ascii output is in a series of columns, values for a single spectrum in each
column. Each column is labelled to show the dataset it has values from and
the spectrum, and group number.
- The column index and dataset name for each column shown in the header of
the file to facilitate column selection for plotting.
- TurboXasScan can now automatically produce ascii files at the end of a scan.
- TurboXas parameters have been updated to include 'write ascii data' field.
- A checkbox has been added to TurboXasExperimentView to control whether Ascii
data is produced or not.
- Unit tests have been added and updated. A new AsciiWriterTest class ensures
that the numbers in each column of ascii output match the corresponding
detector datasets in the Nexus file.

Change-Id: I341d8449a273587720777f0034261ecc400b62f7