bleclipse: improve determination of dls-config/bin & workspace
If bleclipse is run through a symlink, DLS_CONFIG_BIN ends up being set to the
directory containing the symlink, not dls-config/bin.

Canonicalise the path to bleclipse before stripping the last component, so we
get the real location of dls-config/bin.

Likewise, change how GDA_WORKSPACE_PARENT is determined.

Change-Id: Ib7516858511d14a5586b719ad9732316d7c08811