I20_1-205 : TurboXas GUI : Improve error handling when energy-position conversion fails
- TurboXasMotorParameters.getPositionForEnergy() now throws an IllegalArgumentException
with a useful error message if the conversion fails. Unit test added to ensure
exception is thrown if an energy is out of range.
- A check is now made in TurboXas gui when showing motor parameters and
starting a scan for failures in energy to position conversion due to out
of range energies. A warning dialog box displaying the error message
is shown if necessary.

Change-Id: I49cfded114e4aea8bb22d5be3b55635d8a9abb43