I20_1-219 : Use a random colour for first plot in TurboXas/Ede plot view
The first plot added to Ede/TurboXAS scan plot view when running a scan is
now given a random colour. Colours for subsequent plots in the scan are
cycled through from this first colour in the usual way.
EdeScanNode has a new 'getNextColour()' function that returns the RGB colour
that can be used for the colour of next plot.

Change-Id: Ic07ac1592d010e831882d61aa6add666344bba57