[DAQ-1619] Modified Mapping view for XANES scanning
This is based on the same base class as the "normal" mapping view but

- excludes sections not relevant for the XANES scanning script
(e.g. beamline configuration, outer scannables
- includes a XANES parameters view
- has a different "submit scan" button which calls a function
run_xanes_scan(), passing an object containing the relevant parameters.

The functionality of run_xanes_scan() is provided by beamline staff.
For the moment, I have put a skeleton script in:
/dls_sw/i14/scripts/Beamline/Utilities/align
that just prints out the parameters.

Note that this is not integrated with the queueing mechanism used
by "standard" scanning, so you should not submit a XANES scan
while any other scan is running.

Change-Id: Ife3e2e73808bc324b54da327f85f917226d4cc8f