[DAQ-1611] Change PV for ring_mode
Previously, ring_mode monitored the PV SR-CS-RING-01:MODE, but the
value for this when the machine is in its normal state can change:
for example, it recently changed from "VMX" to "DIAD".

With this change, we monitor instead SR-CS-FILL-01:TOPUPMODE, which
will always have the value "Normal" when the machine is
in its normal mode of operation.

Change-Id: I5c851c58dd330bb63715ee79345a03b54b4eef60