I20_1-225 : Add scannable to open/close shutter at start/end of scan
Scannable that simply open a shutter at the start of a scan and closes
it at the end. Shutter is also closed if scan finishes early for some reason.
Unlike ShutterChecker, this scannable simply tries to open/close the shutter,
without complaining if it's not possible e.g. machine day/shutdown or if
hutch door is open.

Change-Id: Ie3b824ef7bfc07a807688b8e3bbc28395150502a