I20_1-225 : Add 'extra scannable' field to TurboXas parameters
- A list of 'extra scannable' has been added to TurboXasParameters.
These scannables are added to the TurboXasScan object in createScan(),
by appending to ScanBase.allScannables list so that the hooks
(aatScanStart, atScanEnd etc) are processed during the scan.
- A list of default 'extra scannables' is stored in the view factory and GUI
and added to the extra scannable list on the parameters object when parameters
are loaded into the GUI.
Unit tests for parameters have been updated to include 'extra scannables'.
Some other minor updates and tidying up has been done as well. 

Change-Id: I5198e14d1943277852a5ab59a1b2a789acb8b320