I20_1-227 : Add method to trigger things during TurboXas scans
New class 'MonitorZebraAndMoveScannable' has been added : this monitors
the number of captured points on a Zebra and calls the moveTo method on a
scannable when a certain number of captures have taken place.
The class maintains a map of spectrum number-position to perform moves
to various different positions depending on how many spectra have been
collected.

Change-Id: I8a9a9208b7c4756f37fd95e8e13cd5b731e53324