P99-118 Restructure P99 config and add new devices
Spring has been deleted and replaced with a minimal form based on the
structure of i08-1.

The non-beamline specific spring imports have all been copied apart
from the machine status since this is not relevant for p99 which has
an independent laser as its beam.

In addition there are changes elsewhere such as removal of Corba.

Device Spring added for: Andor 2 detector, Malcolm and the
J08 stage have been added to dummy and live modes.

Configuration for other devices such as Andor 3 are likely to be added
back as required once they have been recommissioned on the new P99.

Change-Id: Ib26d6297bd6e448ed43aaf3582d7273812ed4151