I20_1-230 : Make EDE detectors reconfigurable
- Reconfigure methods have been added to EdeFrelon and XhDetector classes.
For Xh/XStrip this will close existing connection to DAServer and attempt
to re-establish it. For Frelon it will simply call the configure method again.
- EdeDetectorBase now creates the 'pixels' array in the constructor. This
array depends of on max number of pixels on the detector (which is
constant), and is now available whether detector has been configured or not.
- Spring config. updated so that the DAServer and detector objects are not
configured at server startup.

Change-Id: Ifce9d87956eab7208d9c09241508b0e3e04e83b1