I20_1-230 : Make EDE detectors configurable from GUI
This change is to allow users to the 'configure' the detector selected in the
'Beamline Alignment' view, performing the same steps as during server startup.
This helps avoid errors on server startup due to missing hardware, and
removes need to adjust spring config when different detectors are connected.
- A button has been added to 'Beamline Alignment' gui to allow Xh/XStrip/Frelon
detector to be (re)configured.
- A new 'ConnectDeviceDialog' dialog has been added to configure a detector/device
from the GUI. It displays a progress bar whilst configuration is taking place,
and also shows any error messages that have occurred.

Other GUI updates : 
 - Name of currently selected detector is persisted to the preference store.
 - Values in strip range combo boxes now update when the detector changes.
 - Detector used by 'live view' updates when selected detector changes.
 - Also, removed redundant variables and methods, replaced listeners with lambdas.

Change-Id: I7f27dc40fb032895e9e3a8e3b4c863a52c6dbbac