[DAQ-1299] Define devices that can be used for focus scans (I08, I14 & P99)
The focus scan wizard now requires that each beamline define which
devices (both Malcolm and non-Malcolm) should be shown as options
in the focus scan setup screen.

Before making a Malcolm device available, you should check that the
appropriate axis (the one defined as focusScannableName in the
FocusScanBean) as one of its "axes to move".

Change-Id: Ic00a1923b0a5a5b213592c2c1e445cc1aec66cbd