I20-394 : Open the photon shutter at the start of bragg offset scans
This change ensures that the photon shutter is opened before doing a
bragg offset optimisation scan. Previously, the dark current collection at
the beginning of the offset scan left the shutter open after it had finished.
However, since this measurement is no longer made for these scans it is
now necessary to open the shutter explicitly.

Change-Id: I9a0390639aebe6d097e83d183c40205371910134