I20-396 : Re-implement 'Welcome screen' to select experiment type
The welcome screen that allows new users to select the experiment type
no longer appears. This is based on extending the IntroPart class and
something changed in the recent eclipse target platform updates that
(accidentally?) prevents this mechanism from working.
Re-implementation of this uses IStartup - this class gets called
each time the client starts up and now contains the logic used to display
the welcome screen :
 - An instance scope preference is used to record the selected experiment type.
 - If the preference has not been set a modal dialog box is displayed to allow
 user to select the experiment type. The dialog box is a renamed and
 slightly modified version of the the old I20IntroPart code.
 - If client is '--reset' the experiment type preference is cleared and
 the welcome screen is displayed again when the client starts up.

Change-Id: I8c4a85eb30f44081774bd9d137b6a5d7db5f0b59