[DAQ-1842] Remove energy from XANES-specific mapping sections
The new XANES mapping view will include the standard "outer
scannables" section. This will allow energy (dcm_enrg) to be
set as an outer scannable and to appear in the mscan command
that is generated, so it doesn't need to be duplicated in the XANES
section.

It also means that the XANES-specific status panel can be deleted,
as the standard status panel will pick up dcm_enrg from the
outer scannables section.

Change-Id: I0ef91e57875fe5af9f67fc61621d6fd79576604a