[DAQ-1938] Add ADDetector and overlay for I14 diagnostic camera
The overlay object is initialised for overlay 1 and is intended for
adding cross hairs to the diagnostic camera image. It can, however,
be used to add any of the shapes supported by Epics.

This change also adds a dummy mode function setup_cross_hairs()
that will add add cross hairs to the simulator.

Change-Id: I00e5133438cee95231145e2d945f249a565d07cd