DAQ 1956/2023/2025 Add DiffractionComposite,link to RegionAndPathController
The DiffractionPathComposite makes use of the controller to keep it in step with the existing mapping view Also add definition to K11 mapping
xml file to set default shape to point with corresponding path and add
Summary viewer class.

Also remove dead OSGiServiceRegister declarations and ensure initial
region and path are plotted.

The K11 Client has a mandelbrot detector defined but the server doesn't
have a corresponding device. We will need this for the demo to show the
Point and shoot perspective working.

Change-Id: Ib37d58679004af7b429886ca7bfe9c803cfa90f0