I20_1-266 : Ede scans : Move useFastShutter preference to ExperimentDataModel
The 'useFastShutter' option is common to all types of Ede scan - this
change moves the option from the SingleSpectrum and TimeResolved data models
into ExperimentDataModel which contains all the other common scan parameters.
The useFastShutter option is now persisted to the preference store along with
the rest of the parameters.
Some property change listeners were also updated to use lambdas.

Change-Id: I2eb16b95f2e09843ecea0ba1798c9297880623f0