I20-421 : Fix incorrect initial xml parameters created at startup
This change is to make the startup screen get displayed early in the
eclipse lifecycle so that the XES/XAS preference is set before any XML
files are created. Previously, the ExperimentProjectNature would create
the initial XML files for 'Experiment_1', but this would be done before
the user had selected the XES/XAS experiment type via the startup dialog.

Change-Id: I9fb1c21e884efe9240b5a1622f5b0fb5e5d042d9