MXGDA-3294 Get BeamDataComponent singleton using findLocal
BDC is special; it has a server side and a client side instance which
have no knowledge of each other. Because of this, when you try to find
the "singleton" instance, sometimes the Finder might give you the
non-local instance. This is bad, as RmiAutomatedExporter will try to
export the non-local "singleton" and fail. If we findLocal we always
get the local "singleton" instance.

Change-Id: Ic38e6b891be0c3a9151a0007bdd3651bb7472555