B16-161 (9.12) Remove panel_name references
Many scripts try to use the now removed Plotter object if panel_name is
defined, so make sure it isn't defined. Ususally the default plot is
fine, but if plots need to go to another view this can be overidden,
usually by setting the panel_name_rcp parameter instead of panel_name.

Change-Id: I1688b71924abbeab1d040c94e9a3f18f612ac7c6