I07-198 Revert "BLVII-160 (8.40) - change VersaStat code to read asynchronously"
The threading was introduced to help speed up scans using this device
which uses XML RPC calls that can be slow to return.

Upon using this with GDA 9.12 these scans were failing. The cause
appears to be due to too many RPCs sent without the server finishing
its previous requests.

It is not clear whether the scans would have worked under older versions
or if anything has changed in GDA that could be responsible.

Removing the threading allows scans to run and the slow down
is to be accepted at the moment.

This reverts commit 7e06b6f491be85a7b13fbdab15ac572185be112c.

Change-Id: I49394d9c3e177b59e6ce11c1ea706940efc156bb