I20_1-276 : TurboXas : Move turbo slit to initial scan position before arming the zebra(s)
- The turbo slit move to initial scan position is now done before arming
the zebra(s). This is to avoid accidentally collecting an extra spectrum
before the scan has really started.
- The initial motor move is made using a 1-point trajectory scan.
Sometimes the Epics motor move fails when a trajectory scan left the motor
in a 'bad' state (e.g. from a 'following error'). A 'good' trajectory scan
move normally works in this case, and also makes the Epics motor record
move work correctly again.
- A 'moveWithTrajectoryScan' method has been added to TurboXasScannable.
This moves the turbo slit to a position in the requested time using 1-point
trajectory scan.

Change-Id: If3aff86b9ab014ad626927c66cae6856c1e97c1b