I20_1-253 : Add GUI components for editing a list of scannable positions.
- NumberTableEditor - A JFace table for editing a  table of numbers. Each
column in the table is stored as a list of Doubles. Rows can be added, removed
from the table, and a series of incremental values can also be added by giving
start, stop, step parameters.
- ScannablePositionsComposite - a class that adds GUI elements to control
parameters for scans that take place at multiple positions.
i.e. whether to collect multiple spectra, which scannable is to be moved
and the position of the scannable for each spectrum.

Change-Id: I87c668ee7eb8d01b87cb4a39a945bc9de5ef31f7