I20_1-253 : Updates for EDE scans with multiple It collections
Updates for EDE scans with multiple It collections :
- Pass multiple It positions as a List of Objects. This allows multiple
scannables to be moved for each It spectrum collection by passing a
scannable group as the scannable to be moved.
- Links to each dataset produced by scannables in a group are added
to the Nexus file at the end of the scan.
- TimeResolvedExperimentParameters : add parameters for multiple It collection
and set up the scan using the new parameters in createTimeResolvedExperiment
and createSingleSpectrumScan methods.

Change-Id: I957738783a68f76ed04107db1cca5b24f84e7c51