I20_1-253 : Add GUI components for editing multiple It positions for EDE scan
- GUI elements have been added to allow single spectrum scans to be setup
with multiple positions for the It part of the collection. This has been
added as an optional section in the SingleSpectrumCollectionView (off by default);
a new SingleSpectrumCollectionViewWithMapping view has been added that extends
the SingleSpectrumCollectionView and switches on the extra GUI controls.
- The Experiment perspective now includes SingleSpectrumCollectionViewWithMapping;
Cyclic view has been removed from the perspective.
- The Single spectrum scans are now run from the GUI by creating a
TimeResolvedExperimentParameters object from the current GUI, and passing
the XML string to the server to create and setup the scan bean. This greatly
simplifies the SingleSpectrumCollectionMode.buildScanCommand() method.

Change-Id: Ie3e6b99694beeb2d3bb2402df150c8f7d35cf4cb