New configuration for Malcolm sims
This adds configuration for three simulated malcolm scans: mapping, tomography and load. These talk to epics iocs for motor controls and the tomography scan takes data from an ADSim ioc.

To include these simulated malcolm scans the top level malcolmsims.xml needs to be imported. This then imports, two files: malcolmsims_scans.xml which configures the scans as malcolm devices; and malcolmsims_axes.xml declares the stages used in the scans: simx, simy, simtheta and simload as well as the motors, which point to epics pvs.

There is a comment in malcolmsims.xml which explains how to get and run the external ioc and the malcolm scans, which are required for the scans to work. It should be noted that this requires RedHat 7 to work properly.

Change-Id: Ifb6e059ce2d17663edcf1dc4905f5dbba2bab222