I20_1-278 : TurboXas scan : Compute 'running averages' during scans
Updates to compute 'running averages' of detector data during a scan:
 - A 'namesOfDatasetsToAverage' list has been added to TurboXasParameters -
this is a list of datasets to be averaged when performing a scan. The datasets
are specified using the format <detector name>/<dataset name>.
e.g. 'scaler_for_zebra/I0', 'buffered_xspress3/FF' etc.
 - During the scan the running average of each specified dataset is calculated,
saved in the Nexus file with the rest of the detector data and displayed
in the plot view. The name of the new dataset is <dataset name>_avg.
 - Unit tests have been updated to test that running average values are created
during the scan and have the correct values. Some minor tidying up has
also been done and small changes made to reduce run time of the tests.

Change-Id: I3c5d74627c441eaede0d6bf0336bd06b0004eb90