B16-100 Create B16 scan perspective
This is three columns, the left for the script explorer, shared editor
space and jython terminal. The middle for the detector plot views where
frames are sent during a scan and also the Scan Plot view. The right
column is for the Live Streams.

This is set as the default perspective in the customization ini.

Change-Id: Ib17bb1b8bcd9053866690fff82ef57aae41921ce