I07-210 Start adding xraycam to config
Seems to work as required for now with the xraycamstats object. This
gets a load of PVs from the stat plugin which end up in the scan's
dat file.

The Spring was copied from one of the old flea cameras so might not
be completely suited.

There was some oddity when running scans initially in that they
seemed to be waiting for the detector's cam plugin to finish but this
does not happen for these cameras as they are always acquiring. When
tried after it seemed to work.

Still need to make dummy mode version.

Change-Id: I53b73ddcfba7544f2dfb9cf5901e8fc32cb40774