B16 Fix EpicsPneumaticCallback control PVs for front end devices
The class appends :STA and :CON to the base PV. These PVs do exist
for these front end devices but from the beamline network they are
read only. There is an additional PV with :BLCON which can be written
to from beamlines so use this to set control and status PVs separately.

Change-Id: I0b00edac7111ec38b8ba5bef3d7b859ca30befe9