I20_1-289 : Remove 'Area detector' components from from Zebra for TurboXas scans
This change is to remove references to the 'Area detector' part of the Zebra
for TurboXas scans. This hasn't been used for some time (default is to have
it switched off) and tended to crash the IOC when running fast scans. The
same information on the captured pulses is available through PVs.

Change-Id: I420c6f5455e9ade0d3632ba4fe98a3d4bf03ef14