I07-223 Add coupled scannable for DCM energy
Moving to an energy requires that the pitch and roll of the DCM
crystal are moved to certain positions in addition to whichever
physical motors the energy EPICS motor moves.

These values are provided in a lookup table: dcmLookupTable.txt.
The motors are coupled using InterpolationFunctions (pitch and roll)
and IdentityFunction for the energy.

The new scannable is called simply 'energy'.

Change-Id: Ib03a42e880611259be46e89e04e72361a147831b