I07-224 Add Xmap to beamline
Add NexusXmapFluorescenceDetectorAdapter which allows the Xmap to
function as a Flourescence detector.

Add client view which opens config for this detector.

Add a bunch of plugins to the i07 client feature to allow the view to
work.

Add plugin to manifest (not sure if this is required), this was mostly
copied from B16.

The remaining Spring and parameters are left over from the last time
that I07 had the Xmap so some of them might not be suitable. The plugin
suffix for the HDF5 plugin has been slightly renamed but this is not
actually used for the software step scans that I07 use.

There are also some properties added which the client Fluorescence
detector view uses.

This is the same 'moveable equipment' detector that B16 uses.
The beamline has requested that it behaves like B18's Xmap.

Change-Id: Ieb0a249856204282cfdf932ac1ac05530f4ad986