I20_1-291 : Make single 'Single spectrum' scans run from GUI stoppable
'Single spectrum' scans run the GUI controls were not started in a separate
thread on the server which meant they were unstoppable stoppable using
the 'stop' button in the client. This change fixes the problem :
- 'Single spectrum' GUI controls now start scans in a background thread
on the server.
- Scans that do 'multiple It' collections now check to see if the thread
has been interrupted or early finish request has been made at start of
each new collection position.
- The 'Single spectrum' scan 'running' status is now determined using a
listener to the scan events.
- Filename is no longer used in SingleSpectrumCollectionModel and has been
removed (it was previously used to detect when a scan finished).

Scans can now be stopped 'nicely' using 'smart stop', or 'fast forward'
buttons, or aborted using the 'stop' button.


Change-Id: I81529936213899d464f13b6d48bb407993f44fd9