I20-1 : Ede plot view - detect Xspress3 group in Nexus file using 'FF_sumI0'
In the Ede plot view, data for Xspress3 is found by looking for a Nexus
group containing a dataset called 'MCAs' which is a link to the MCA data
in the xspress3 hdf file. However, sometimes the link is missing or MCA writing
is switched off. It is better to look for FF_sumI0 instead, as this dataset
is always present.

Change-Id: Ic26f4272d8ea0f522e5f53f73739ba3776fcd470