B16-184 Use display.asyncExec() to initialise plot views
The previous code worked fine when the perspective was not set as the
default i.e. a different perspective loads when the client starts.

However when using this as the default it caused some very strange
behaviour which lead to the views being permanently open on the RHS
of all perspectives. They were also glitchy when clicked on.

Presumably this happened because this code is called as the client is
starting up and certain components might not be initialised properly.

Switching to asyncExec() appears to fix the problem.

Change-Id: If7bcf319ea82a0e46d6eb89c81264da2979210e4