DAQ-2360 - Delete empty beans to prevent NPEs in BatonManager
I08 and I14 have an "empty" bean as the first list entry in their
GDAMetada objects. In previous versions of GDA this was fine, as the
empty bean becomes a single whitespace character entry in the list due
to whatever reflection magic creates this object.

In current master (and 9.14) the reflection magic interprets the empty
bean as null, so when equals gets called on the first entry an NPE is
thrown.

These empty beans don't seem to serve any purpose so I deleted them, now
the visit gets set correctly.

Update - This bug was created by 21221

Change-Id: Iaaf2827577d68054376cbf1ba4d619b01fd135f3