I20 : Updated DetectorPreparer after Medipix plugin changes
For XES scans, the Medipix now uses an 'additional plugin' list containing
a 'mutable ROI'. This considerably simplifies setting the ROI on the detector
at the start of scans. The original plugin list is restored at the end of
the the scan.
The plugin list and 'mutable roi' are both injected into the detector
preparer in localStation.py

Change-Id: I4bcc80cc23f45497552d2d3d05eb9e39dbd9f58f