I07-222 Switch to :MRREADBACK for qbpm ranges
The :RANGE PV is assosicated with the :SETRANGE PV which gives
a feedback once the request to change has been completed by the
device. This means that if it is set to 'Auto' both of these PVs
will be 'Auto'.

:MRREADBACK will give the actual measuring range i.e. even if
set to auto this will give the actual current the device is using.

Change-Id: Ie8358d5e0d557bf4fb853e8e53dc77f722621080