B18-435 : Changes to localStation.py to improve exception handling
Parts of localStation that setup xspress3, xspress4 detectors can cause
the script to exit early due to exception being thrown (due to missing PVs
when the IOCs are not running).
This change refactors xspress3, xspress4 and mythen setup code to
separate functions which are run in a try-catch blocks, so that exceptions
do not prevent localStation from completing.

Change-Id: I82e0fbed86d2ea1daef61758ff2338eb5417af76