I08-269 Initialise andor_adbase with initialBinX/Y rather than BinX/Y
Setting BinX/Y makes an immediate call to the EPICS controller.
If the Andor camera is unavailable, this call will throw an exception
and the GDA server will fail to start.

Setting initialBinX/Y just sets a variable in the object, which is later
(at configure time) used to set the value in EPICS. This will still
fail, but failure in the configure phase just logs an error, and
server startup will continue.

Change-Id: I6cd42273d1e5944ed40bd7d232017e293e216384