I20-469 : Make "XES calibration view" GUI work for integer energies
This fixes a problem where the offsets in "XES calibration GUI" were not
calculated when passed an integer energy. This was caused by a cast
exception (the long int returned by NumberFormat could not be cast to Double)
which resulted in the offset calculation not getting run, with no useful
errors in the log file.

Change-Id: I806309f6151a8360c438ef37b2ec739df10767e6