Remove loadjava.sh shebang
The purpose of the loadjava.sh file is to modify the shell's environment
so that the correct Java VM will be used.  The loadjava.sh file will
only be sourced from another shell, and all existing users of this file
source it.  Since executing it as a shell script would create a new
environment and then immediately throw the environment away when the
script exits, executing it doesn't make sense, so remove the /bin/bash
shebang.

Change-Id: I862547c2d9495cac57448acb8475dfd9fe213cf5