B16-208 Add 'wait for condition' scannable
This was used when a detector that was brought in by users hence was
not integrated in Epics but rather could be sent a trigger to begin
acquisition and itself would output a signal whilst it was acquiring.
This signal was connected to a zebra.

These scannanles offer a crude control for running scans with detectors
like this. It could be improved by combining into a single scannable
since they are dependent on each other.

Add minimal logging to the condition class.

Change-Id: Id5fcc3aecdeb88b15a206d4e68761cbebf3c8c46