I07 (9.15) Delete temporary scannable used in metadata loop
Running this loop logs lots of warnings about overwriting scannables.
As part of this logging toString is called on the scannable in question.
This should be fine but many of the Jython scannables on I07 perform
getPosition in their toString methods which causes an exception to be
thrown if the device is not available. This can cause the logging
appender to crash.

Change-Id: I22e51be2626bd5ef2a3cdd78c569b23224e1bd14