B16-227 Add function to create subdirectory within visit
This is required by users who are not principally using GDA for data
collection hence they have to way of creating subdirectories in the
visit due to the permissions as outlined at:
https://confluence.diamond.ac.uk/x/p4vtAw

The detector user will be writing files but the Odin framework does
not support creating directories, i.e. expects them to have been created
by GDA (or other means).

Change-Id: Ife65fb8f740d87311495d084dbaf9424c8f094e0