I07-265 Extract little used devices into separate xml
These are rarely used devices so they can be imported when required
by uncommenting them in ${config}/servers/main/live/transient.xml.

This is somewhat a temporary measure and this facility should be replaced
with proper transient device support once it is included in GDA.

Note this also adds a ScannableGroup for CLASP.

The devices concerned are: dcdx, gw stage, clasp and p2m stage.

Change-Id: Ife331aef2c590b311032a34aa835e6d561c32f1c