MXGDA-3388 Make IBatonManager.assignBaton() aware of baton passer
Currently, a baton pass action can only be performed instantly. In order
to be able to queue up a baton pass, we need to be able to check at the
time of the baton pass queue action if the "passing client" is _still_
the baton holder. If not reject the pass - we don't want the queue to
be able to take the baton from another user.

Change-Id: I1adedef0a6779490495292d56ac6cce56a8c2dbf