DAQ-2666 Fix output path calculation for tomography scan
Fix the calculation of subdirectory for temporary files.
The previous algorithm assumed 6 elements in the path, which is not
necessarily true (and is not true in I13 dummy mode).
This version removes the "real" subdirectory suffix (as set in
gda.data.scan.datawriter.datadir.subdir.live) and replaces it with tmp.

A corresponding change has been made to ScanParameterDialog

Change-Id: I7b8f5a3fd04a63ff2b9e96185842fbd2d2cf19e8