B18 : Make Ln2Scannable use IScannableGroup for 'cryo' scannable
Ln2Scannable was modified as part of change abe1a6c46 to use IScannableGroupNamed
interface for 'cryo' scannable group, but corresponding spring config was
not also changed, causing server startup to fail. This change makes 'cryo'
use the IScannableGroup interface again and changes the 'Move complete'
message in moveTo to just display the position of the 'cryo' scannable.

Change-Id: Ia2b832130e17f5596b3790b83996f742e37f63aa