I08-317: I08-1: add scannables for OSA & combined ZPZ/OSA scannable
See Jira ticket for details of the requirement.

This change adds:
- configuration for the OSA motors
- a SequentialScannable zpz_osa that combines the motion of
ZonePlateZ and the OSA as per the requirements in the ticket
- energyFocus scannables, similar to those on i08, but using zpz_osa
instead of ZonePlateZ

Changes as a consequence of this:
- energyFocus.py renamed to initialiseEnergyFocusFunction.py to make
it clearer what it does
- energyFocus.xml moved from shared i08-shared to i08-config, as
the i08-1 version needs to be different
- corrected the limits on the dummy zone plate motors to be the
same as the real motors

NB: The function that couples the motion of ZPZ and OSAZ is
initialised at server startup on the assumption that the ZPZ and OSAZ
are in the correct relative positions.
If this is not the case, then initialise_zpz_osa_function() must be
run again (or the namespace must be reset) when the two devices
are in a consistent position.

Change-Id: I7a4cbc54e84101243f62ec12b6ef72376a7a5d32