I16-470 (9.15) Add rocker scannable class
RockerScannable rocks an axis during exposures. It can be configured with
the delta to rock by and the number of rock cycles per exposure. In a
scan it will calculate the speed needed to to achieve the required rocks
given the exposure time given to the scannable, starting at the current
position and returning to the original position.

Since rocks are relative, the absolute positions rocked over can change
throughout the scan, unlike ContinuouslyRockingScannable, based on the
positions the motors are set to by scannables with a lower level. For
instance you can rock phi, even though phi may be one of the motors set
by the hkl positions you are scanning over.

Change-Id: I31f699d876b12ab8b6e4aefc5fb3975c5ec71379