B16-240 Remove GDA_DEBUG from startup script
This causes the client process to bind to the debug port. If an
additional client is launched on the same machine (which will be by a
different user) it fails because the port is already in use.

Debug shouldn't be the default and can be enabled when required using
the --debug command line option to gda client.

Change-Id: I1df97bad2b01cc33c967cec7fee425da32575c93