K11-603 Apply experiment structure to Point and Shoot
This change makes the output of a Point and Shoot session conform to the
concept of the experiment, and encapsulates all the individual scans in
the session in a single session file. This file is at the same level of
a standalone acquisition with an experiment.

Example output structure:

{datadir}/MyExperiment
		|- MyExperiment.nxs  		// <- index file
		|- SimpleTomography/
			|-SimpleTomography.nxs
		|- SimpleDiffraction/
			|-SimpleDiffraction.nxs
		|- MyPointAndShoot
			|-MyPointAndShoot.nxs   // <- P&S session file
			|-MyPointAndShoot1/
				|-MyPointAndShoot1.nxs
			| ...
			|-MyPointAndShootn/
				|-MyPointAndShootn.nxs

What was the CtrlClickToScanHandler OSGi event handler is now part of
PointAndShootController, which is instantiated when a session starts and
manually registers the OSGi event handler.

Change-Id: Id7173281b98b8ca8bdcc52d84592a2615e4a882c