i18: Reorganise configuration
It was difficult to find and maintain individual bits of configuration
when most was under a single devices.xml or scannables.xml which had the
entire beamline description.

Now beans are put under their device categories e.g. slits, mirrors,
and each category in _common has an index file which the common server
imports. This file is responsible for importing its mode-specific
dependencies and then the common components if applicable.

Change-Id: Ib77a320a94bd7b7a8d5dec2042b3151a8dff3557