DAQ-2841 Skip tests using GDALargeTestFiles if system property not set
These tests use files stored at /dls_sw/dasc/GDALargeTestFiles/ (or
a mirror of this on a Jenkins agent's local disk). This needs to be
set i.e. -DGDALargeTestFilesLocation=/path/to/GDALargeTestFiles for the
tests to run.

Change so that tests are skipped (Using Junit's Assume) rather than
fail so that they don't have to be manually excluded in the pom.

Change-Id: Ie03bc887d03ee7d6ead30b0d5283c44297ededa8