I08-342 Update dummy configurations
Now that there is a "real Malcolm/dummy motors" configuration
available, we do not need the  DummyMalcolmDevice configurations.

Move the DeviceInformation beans to the live malcolm.xml and delete
the dummy malcolm.xml.

Also remove the (commented-out) Mandelbrot configuration and other
example detectors, as there are newer versions in dls-config
which can be included if necessary.

Change-Id: Ib1b9e81dbae526e8e167597e5fb426b479c6ee2d